Matching service

+++ src/main/java/com/volvic20/Volvic20/models/GoogleAPI/payload/ArrivalLocation.java
@@ -1,27 +1,27 @@
 package com.volvic20.Volvic20.models.GoogleAPI.payload;
 
 public class ArrivalLocation {
-    private double latitude;
-    private double longitude;
+    private float latitude;
+    private float longitude;
 
-    public ArrivalLocation(double latitude, double longitude) {
+    public ArrivalLocation(float latitude, float longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(double latitude) {
+    public void setLatitude(float latitude) {
         this.latitude = latitude;
     }
 
-    public double getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(double longitude) {
+    public void setLongitude(float longitude) {
         this.longitude = longitude;
     }
 }
